refactor: Start decomposing CommunityCommunity
